--- src/modules/storage.js
@@ -1,8 +1,0 @@
-export const getTodosFromStorage = () => {
-  const todos = localStorage.getItem('todos');
-  return todos ? JSON.parse(todos) : [];
-};
-
-export const updateTodosInStorage = (todos) => {
-  localStorage.setItem('todos', JSON.stringify(todos));
-};
